Check if the register_user form is likely a resent message.
The current timeout is quite short (10 seconds).  I routinely
run into this in my current setup.  The timeout should be raised,
but there is still the issue that what happens if the timeout
happens on the client side but the operation is successfully
carried out to the end?  After this, if a device already has
a primary user, it matches the submitted user, and user password
matches the user account, we return success (the user information
is not updated in any way).